+++ Assets/Scenes/CRPrototypeLevel.unity
@@ -355,7 +355,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!65 &334756907
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -1339,7 +1339,7 @@ Transform:
   m_GameObject: {fileID: 1150559817}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 2, z: -13.36}
+  m_LocalPosition: {x: 0, y: 1.65, z: -13.36}
   m_LocalScale: {x: 0.9002769, y: 2.8383932, z: 0.9002769}
   m_ConstrainProportionsScale: 0
   m_Children:
