--- Assets/Scripts/Collectables/ScoreControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e905552493412fe40ba642499cf8f0f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
