+++ Assets/Scripts/PCG/LanePopulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6ab1bc0b0a67ed4dbfdc8ad87d86511
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
