+++ Assets/Scripts/PCG/SceneryPopulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5705958c7fc0b2843b9b850962361f95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
