+++ ProjectSettings/TagManager.asset
@@ -15,6 +15,7 @@ TagManager:
   - Tree4
   - Rock
   - House
+  - Skip
   layers:
   - Default
   - TransparentFX
